refactor: reorganize permission views
