--- frontend/src/views/PermissionConfig.vue
@@ -1,10 +1,0 @@
-<template>
-  <PermissionTreeTab />
-</template>
-
-<script setup>
-import PermissionTreeTab from './permission/PermissionTreeTab.vue'
-</script>
-
-<style scoped>
-</style>
